# api/openapi.yaml
openapi: 3.0.4
info:
  title: Dployr API
  description: |
    Your app, your server, your rules!
    
    ## Authentication
    Dployr instances use JWT-based authentication with tokens **issued by base**.

    ### Token Types (issued by base)
    - **Access Token**: Short-lived (e.g. 10 minutes) for API requests to instances
    - **Refresh Token**: Long-lived (24+ hours) for obtaining new access tokens from base

    ### Authentication Flow
    See https://docs.dployr.dev/auth for more details.
    
    ## Authorization
    Role-based access control with four levels:
    - **Owner**: Full system control, can manage admins
    - **Admin**: Infrastructure management, user management, secrets
    - **Developer**: Deploy apps, view logs and events
    - **Viewer**: Read-only access to services

    See https://docs.dployr.dev/auth for more details.
  version: 1.1.1
  contact:
    name: Emmanuel Madehin
    email: hello@dployr.dev
  license:
    name: MIT
    url: https://raw.githubusercontent.com/dployr-io/dployr/refs/heads/master/LICENSE
servers:
  - url: "{address}"
    description: Configurable server
    variables:
      address:
        default: http://localhost:7879
        description: Instance URL
security:
  - BearerAuth: []
paths:
  /deployments:
    get:
      tags:
        - Deployments
      summary: List deployments
      description: Retrieve a list of deployments (Developer+ required)
      operationId: listDeployments
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          description: Maximum number of deployments to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Number of deployments to skip
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: List of deployments
          content:
            application/json:
              schema:
                type: object
                properties:
                  deployments:
                    type: array
                    items:
                      $ref: '#/components/schemas/Deployment'
                  total:
                    type: integer
                    description: Total number of deployments
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Deployments
      summary: Create deployment
      description: Create a new deployment (Developer+ required)
      operationId: createDeployment
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeployRequest'
      responses:
        '201':
          description: Deployment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeployResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /services:
    get:
      tags:
        - Services
      summary: List services
      description: Retrieve a list of services (Viewer+ required)
      operationId: listServices
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          description: Maximum number of services to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Number of services to skip
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: List of services
          content:
            application/json:
              schema:
                type: object
                properties:
                  services:
                    type: array
                    items:
                      $ref: '#/components/schemas/Service'
                  total:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /services/{serviceId}:
    get:
      tags:
        - Services
      summary: Get service
      description: Retrieve a specific service by ID (Viewer+ required)
      operationId: getService
      security:
        - BearerAuth: []
      parameters:
        - name: serviceId
          in: path
          required: true
          description: Service ID
          schema:
            type: string
      responses:
        '200':
          description: Service details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Services
      summary: Update service
      description: Update a service configuration (Developer+ required)
      operationId: updateService
      security:
        - BearerAuth: []
      parameters:
        - name: serviceId
          in: path
          required: true
          description: Service ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceUpdate'
      responses:
        '200':
          description: Service updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    patch:
      tags:
        - Services
      summary: Partially update service
      description: Partially update a service configuration (Developer+ required)
      operationId: patchService
      security:
        - BearerAuth: []
      parameters:
        - name: serviceId
          in: path
          required: true
          description: Service ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceUpdate'
      responses:
        '200':
          description: Service updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Services
      summary: Delete service
      description: Delete a service (Admin+ required)
      operationId: deleteService
      security:
        - BearerAuth: []
      parameters:
        - name: serviceId
          in: path
          required: true
          description: Service ID
          schema:
            type: string
      responses:
        '204':
          description: Service deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /logs/stream:
    get:
      tags:
        - Logs
      summary: Open log stream
      description: |
        Open a keep alive connection to stream logs via server sent events (SSE) (Developer+ required).
        
        - `data:` events contain log lines
        - `event: error` indicates streaming errors
        - `event: done` indicates stream completion
        - `: heartbeat` every 15 seconds to keep the connection alive
      operationId: openLogStream
      security:
        - BearerAuth: []
      parameters:
        - name: token
          in: query
          required: true
          description: Bearer token for authentication
          schema:
            type: string
        - name: id
          in: query
          required: true
          description: Service or deployment ID to stream logs for
          schema:
            type: string
      responses:
        '200':
          description: Logs stream is successfully established
          content:
            text/event-stream:
              schema:
                type: string
                description: Server-Sent Events stream with log data
                example: |
                  data: cloning repository
                  
                  data: installing nodejs@18.1.0
                  
                  : heartbeat
                  
                  event: done
                  data: stream completed
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '405':
          description: Method not allowed (only GET is supported)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /proxy/status:
    get:
      tags:
        - Proxy
      summary: Get proxy status
      description: Get the current status of the proxy service (Developer+ required)
      operationId: getProxyStatus
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Proxy status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProxyStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /proxy/restart:
    post:
      tags:
        - Proxy
      summary: Restart proxy
      description: Restart the proxy service (Admin+ required)
      operationId: restartProxy
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Proxy restarted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Proxy restarted successfully"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /proxy/add:
    post:
      tags:
        - Proxy
      summary: Add proxy route
      description: Add a new route to the proxy configuration (Admin+ required)
      operationId: addProxyRoute
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProxyRoute'
      responses:
        '201':
          description: Route added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Route added successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /proxy/remove:
    delete:
      tags:
        - Proxy
      summary: Remove proxy route
      description: Remove a route from the proxy configuration (Admin+ required)
      operationId: removeProxyRoute
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Route removed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Route removed successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /system/status:
    get:
      tags:
        - System
      summary: Get system status
      description: Get overall system health and status (Viewer+ required)
      operationId: getSystemStatus
      security:
        - BearerAuth: []
      responses:
        '200':
          description: System status information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /system/info:
    get:
      tags:
        - System
      summary: Get system information
      description: Get system version and configuration info (Admin+ required)
      operationId: getSystemInfo
      security:
        - BearerAuth: []
      responses:
        '200':
          description: System information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemInfo'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /system/doctor:
    post:
      tags:
        - System
      summary: Run system doctor
      description: Run system diagnostics and attempt auto-fixes for core dependencies.
      operationId: runSystemDoctor
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Doctor completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemDoctorResult'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /system/install:
    post:
      tags:
        - System
      summary: Install or upgrade dployr
      description: Install a specific version of dployr or the latest version, then run the system doctor.
      operationId: installSystem
      security:
        - BearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstallRequest'
      responses:
        '200':
          description: Install and doctor completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemDoctorResult'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /system/register:
    post:
      tags:
        - System
      summary: Register instance with base
      description: |
        Register this dployr instance with the dployr base.

        This endpoint is typically called by base with a signed claim token and
        instance metadata. It does **not** require authentication but validates
        the claim against the locally stored token obtained from base during
        installation.
      operationId: registerInstance
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterInstanceRequest'
      responses:
        '200':
          description: Instance successfully registered
        '500':
          $ref: '#/components/responses/InternalServerError'
  /system/domain:
    post:
      tags:
        - System
      summary: Request domain from base
      description: |
        Request and assign a managed domain for this dployr instance from the base control plane.

        This endpoint is typically called during installation with a token issued by base and
        the instance's address. It does **not** require authentication.
      operationId: requestDomain
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestDomainRequest'
      responses:
        '200':
          description: Domain successfully requested and instance registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestDomainResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RuntimeObj:
      type: object
      required:
        - type
      properties:
        type:
          $ref: '#/components/schemas/Runtime'
        version:
          type: string
          example: "18.0.0"
    RemoteObj:
      type: object
      properties:
        url:
          type: string
          format: uri
          example: "https://github.com/user/repo.git"
        branch:
          type: string
          example: "main"
        commit_hash:
          type: string
          example: "abc123def456"
...
    DeploymentStatus:
      type: string
      enum:
        - pending
        - in_progress
        - failed
        - completed
    Deployment:
      type: object
      properties:
        id:
          type: string
          example: "01HN2K3M4N5P6Q7R8S9T0U1V2W"
        user_id:
          type: string
          description: ID of the user who created the deployment
        user_email:
          type: string
          format: email
          description: Email of the user who created the deployment
        config:
          $ref: '#/components/schemas/Blueprint'
        status:
          $ref: '#/components/schemas/DeploymentStatus'
        metadata:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    DeployRequest:
      type: object
      required:
        - name
        - user_id
        - source
        - runtime
      properties:
        name:
          type: string
          example: "my-app"
        description:
          type: string
          example: "My application description"
        user_id:
          type: string
        source:
          type: string
          enum: [remote, image]
        runtime:
          $ref: '#/components/schemas/RuntimeObj'
        version:
          type: string
        run_cmd:
          type: string
          example: "npm start"
        build_cmd:
          type: string
          example: "npm run build"
        port:
          type: integer
          minimum: 1
          maximum: 65535
          example: 3000
        working_dir:
          type: string
          example: "/app"
        static_dir:
          type: string
          example: "/app/dist"
        image:
          type: string
          example: "node:18-alpine"
        env_vars:
          type: object
          additionalProperties:
            type: string
        secrets:
          type: object
          additionalProperties:
            type: string
        remote:
          $ref: '#/components/schemas/RemoteObj'
        domain:
          type: string
          example: "myapp.example.com"
        dns_provider:
          type: string
          example: "cloudflare"
    DeployResponse:
      type: object
      properties:
        success:
          type: boolean
        id:
          type: string
          example: "01HN2K3M4N5P6Q7R8S9T0U1V2W"
        name:
          type: string
          example: "my-app"
        created_at:
          type: string
          format: date-time
    Service:
      type: object
      properties:
        id:
          type: string
          example: "01HN2K3M4N5P6Q7R8S9T0U1V2W"
        name:
          type: string
          example: "my-service"
        description:
          type: string
          example: "My service description"
        source:
          type: string
        runtime:
          $ref: '#/components/schemas/Runtime'
        runtime_version:
          type: string
          example: "18.0.0"
        run_cmd:
          type: string
          example: "npm start"
        build_cmd:
          type: string
          example: "npm run build"
        port:
          type: integer
          example: 3000
        working_dir:
          type: string
          example: "/app"
        static_dir:
          type: string
          example: "/app/dist"
        image:
          type: string
          example: "node:18-alpine"
        env_vars:
          type: object
          additionalProperties:
            type: string
        status:
          type: string
        remote:
          type: string
          format: uri
        branch:
          type: string
          example: "main"
        commit_hash:
          type: string
          example: "abc123def456"
        blueprint:
          $ref: '#/components/schemas/Blueprint'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ServiceUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        runtime_version:
          type: string
        run_cmd:
          type: string
        build_cmd:
          type: string
        port:
          type: integer
          minimum: 1
          maximum: 65535
        working_dir:
          type: string
        static_dir:
          type: string
        image:
          type: string
        env_vars:
          type: object
          additionalProperties:
            type: string
        branch:
          type: string
    ProxyStatus:
      type: object
      properties:
        status:
          type: string
          example: "running"
    ProxyRoute:
      type: object
      required:
        - domain
        - upstream
      properties:
        domain:
          type: string
          example: "example.com"
        upstream:
          type: string
          example: "http://localhost:3000"
    SystemStatus:
      type: object
      properties:
        status:
          type: string
          enum: [healthy, degraded, unhealthy]
          example: "healthy"
        uptime:
          type: string
          description: System uptime duration
          example: "72h30m15s"
        services:
          type: object
          properties:
            total:
              type: integer
              example: 5
            running:
              type: integer
              example: 4
            stopped:
              type: integer
              example: 1
        proxy:
          type: object
          properties:
            status:
              type: string
              example: "running"
            routes:
              type: integer
              example: 3
    SystemInfo:
      type: object
      properties:
        build:
          type: object
          properties:
            version:
              type: string
              example: "v0.1.1-beta.17"
            commit:
              type: string
              example: "a1b2c3d4"
            build_date:
              type: string
              description: Build date - bundled at build time
            go_version:
              type: string
              example: "go1.21.0"
        hardware:
          type: object
          properties:
            os:
              type: string
              example: "linux"
            arch:
              type: string
              example: "amd64"
            cpu_count:
              type: integer
              example: 8
            hostname:
              type: string
              nullable: true
            kernel:
              type: string
              nullable: true
              example: "Linux 6.8.0-35-generic"
            mem_total:
              type: string
              nullable: true
              example: "7.5Gi"
            mem_used:
              type: string
              nullable: true
            mem_free:
              type: string
              nullable: true
            swap_total:
              type: string
              nullable: true
            swap_used:
              type: string
              nullable: true
        storage:
          type: object
          properties:
            partitions:
              type: array
              items:
                type: object
                properties:
                  filesystem:
                    type: string
                  size:
                    type: string
                  used:
                    type: string
                  available:
                    type: string
                  use_percent:
                    type: string
                  mountpoint:
                    type: string
            devices:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
                  size:
                    type: string
                  type:
                    type: string
                  mountpoints:
                    type: array
                    items:
                      type: string
    SystemDoctorResult:
      type: object
      properties:
        status:
          type: string
          example: "ok"
        output:
          type: string
          description: Combined textual output from the doctor/install scripts
        error:
          type: string
          description: Error message if status is "error"
    RegisterInstanceRequest:
      type: object
      required:
        - claim
        - instance_id
      properties:
        claim:
          type: string
          description: Signed claim token issued by base
          example: "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
        instance_id:
          type: string
          description: Unique identifier assigned to this instance by base
          example: "inst_01HXYZABCDEF123456"
        issuer:
          type: string
          description: Expected token issuer (for validation)
          example: "https://base.dployr.dev"
        audience:
          type: string
          description: Expected token audience (for validation)
          example: "dployr-instance"
    RequestDomainRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          description: Installation token issued by base for this instance
    InstallRequest:
      type: object
      properties:
        version:
          type: string
          description: dployr version tag to install (e.g. v0.1.1-beta.17). If omitted, installs latest.
    RequestDomainResponse:
      type: object
      properties:
        domain:
          type: string
          description: Assigned domain for this instance
          example: "my-awesome-app.dployr.dev"
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: object
          description: Additional error details
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Unauthorized"
            code: "auth.unauthorized"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Insufficient permissions"
            code: "auth.forbidden"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Resource not found"
            code: "resource.not_found"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Internal server error"
            code: "runtime.internal_server_error"
...